+++ kgcnn/losses/__init__.py
@@ -1,0 +1,9 @@
+from ._loss import (
+    RaggedMeanAbsoluteError,
+    BinaryCrossentropyNoNaN
+)
+
+__all__ = [
+    "RaggedMeanAbsoluteError",
+    "BinaryCrossentropyNoNaN"
+]